Added a dark mode
This was a bit of work but it was pretty fun to do.

The best part was getting CoPilot to autocomplete all the boilerplate code! I saved so many keystrokes that way.

I feel like AI is going to be really good for pair programming and autocompleting tasks... Probably like having your own intern/junior dev?